requirements.txt
PyMuPDF==1.25.5
Kivy==2.3.0
plyer==2.1.0
sentence-transformers==4.0.2
openai==1.70.0
polars==0.20.21
pandas==2.1.4
scikit-learn==1.4.1.post1
numpy==1.26.3
google-genai==1.12.1
protobuf~=4.25.3
seaborn~=0.13.2
matplotlib~=3.8.3
statsmodels~=0.14.2
scipy~=1.12.0